refactor(llc): move the channel files into their own directory
