+++ packages/stream_chat/lib/src/client/channel/channel_read_helper.dart
@@ -1,4 +1,4 @@
-import '../../stream_chat.dart';
+import '../../../stream_chat.dart';
 
 /// Extension methods for reading related operations on a ChannelClientState.
 extension ChannelReadHelper on ChannelClientState {
