+++ packages/stream_chat/test/src/client/channel/channel_read_helper_test.dart
@@ -2,8 +2,8 @@ import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
-import '../fakes.dart';
-import '../mocks.dart';
+import '../../fakes.dart';
+import '../../mocks.dart';
 
 void main() {
   group('ChannelReadHelper', () {
