+++ packages/stream_chat/test/src/client/channel/channel_test.dart
@@ -4,9 +4,9 @@ import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
-import '../fakes.dart';
-import '../matchers.dart';
-import '../mocks.dart';
+import '../../fakes.dart';
+import '../../matchers.dart';
+import '../../mocks.dart';
 
 void main() {
   ChannelState _generateChannelState(
